Revised implementation of reporting of external impacts
