+++ utilities/ImpactUtilities.groovy
@@ -13,6 +13,7 @@ import java.util.regex.*
 @Field BuildProperties props = BuildProperties.getInstance()
 @Field def gitUtils= loadScript(new File("GitUtilities.groovy"))
 @Field def buildUtils= loadScript(new File("BuildUtilities.groovy"))
+@Field def reportingUtils= loadScript(new File("ReportingUtilities.groovy"))
 @Field String hashPrefix = ':githash:'
 @Field def resolverUtils
 
@@ -181,13 +182,13 @@ def createImpactBuildList(RepositoryClient repositoryClient) {
 	// Perform analysis and build report of external impacts
 	if (props.reportExternalImpacts && props.reportExternalImpacts.toBoolean()){
 		if (props.verbose) println "*** Analyze and report external impacted files."
-		reportExternalImpacts(repositoryClient, changedFiles)
+		reportingUtils.reportExternalImpacts(repositoryClient, changedFiles)
 	}
 
 	// Document and validate concurrent changes
 	if (props.reportConcurrentChanges && props.reportConcurrentChanges.toBoolean()){
 		if (props.verbose) println "*** Calculate and document concurrent changes."
-		calculateConcurrentChanges(repositoryClient, buildSet)
+		reportingUtils.calculateConcurrentChanges(repositoryClient, buildSet)
 	}
 
 	return [buildSet, deletedFiles]
@@ -223,11 +224,17 @@ def createMergeBuildList(RepositoryClient repositoryClient){
 			if (props.verbose) println "** Found build script mapping for $changedFile. Adding to build list"
 		}
 	}
+	
+	// Perform analysis and build report of external impacts
+	if (props.reportExternalImpacts && props.reportExternalImpacts.toBoolean()){
+		if (props.verbose) println "*** Analyze and report external impacted files."
+		reportingUtils.reportExternalImpacts(repositoryClient, changedFiles)
+	}
 
 	// Document and validate concurrent changes
 	if (props.reportConcurrentChanges && props.reportConcurrentChanges.toBoolean()){
 		if (props.verbose) println "*** Calculate and document concurrent changes."
-		calculateConcurrentChanges(repositoryClient, buildSet)
+		reportingUtils.calculateConcurrentChanges(repositoryClient, buildSet)
 	}
 
 	return [buildSet, deletedFiles]
@@ -546,198 +553,6 @@ def scanOnlyStaticDependencies(List buildList, RepositoryClient repositoryClient
 				}
 			} else {
 				if (props.verbose) println ("*** Skipped scanning outputs of $buildFile. No language prefix found.")
-			}
-		}
-	}
-}
-
-/*
- * Method to query the DBB collections with a list of changed files  
- * Configured through reportExternalImpacts* build properties
- */
-
-def reportExternalImpacts(RepositoryClient repositoryClient, Set<String> changedFiles){
-	// query external collections to produce externalImpactList
-
-	Map<String,HashSet> collectionImpactsSetMap = new HashMap<String,HashSet>() // <collection><List impactRecords>
-	List<Pattern> collectionMatcherPatterns = createMatcherPatterns(props.reportExternalImpactsCollectionPatterns)
-
-	// caluclated and collect external impacts
-	changedFiles.each{ changedFile ->
-
-		List<PathMatcher> fileMatchers = createPathMatcherPattern(props.reportExternalImpactsAnalysisFileFilter)
-
-		if(matches(changedFile, fileMatchers)){
-
-			if (props.reportExternalImpactsAnalysisDepths == "simple"){
-				// Simple resolution without recursive resolution
-				String memberName = CopyToPDS.createMemberName(changedFile)
-
-				def ldepFile = new LogicalDependency(memberName, null, null);
-				repositoryClient.getAllCollections().each{ collection ->
-					String cName = collection.getName()
-					if(matchesPattern(cName,collectionMatcherPatterns)){ // find matching collection names
-						if (cName != props.applicationCollectionName && cName != props.applicationOutputsCollectionName){
-							def Set<String> externalImpactList = collectionImpactsSetMap.get(cName) ?: new HashSet<String>()
-							def logicalFiles = repositoryClient.getAllLogicalFiles(cName, ldepFile);
-							logicalFiles.each{ logicalFile ->
-								def impactRecord = "${logicalFile.getLname()} \t ${logicalFile.getFile()} \t ${cName}"
-								// if (props.verbose) println("*** $impactRecord")
-								externalImpactList.add(impactRecord)
-							}
-							collectionImpactsSetMap.put(cName, externalImpactList)
-						}
-					}
-					else{
-						//if (props.verbose) println("$cName does not match pattern: $collectionMatcherPatterns")
-					}
-				}
-			}
-			else if(props.reportExternalImpactsAnalysisDepths == "deep"){
-				// Recursive analysis to support nested scenarios
-
-				// Configure impact resolver
-				ImpactResolver impactResolver = new ImpactResolver().file(changedFile).repositoryClient(repositoryClient)
-
-				String impactResolutionRules = props.getFileProperty('impactResolutionRules', changedFile)
-				impactResolver.setResolutionRules(buildUtils.parseResolutionRules(impactResolutionRules))
-
-				repositoryClient.getAllCollections().each{ collection ->
-					String cName = collection.getName()
-					if(matchesPattern(cName,collectionMatcherPatterns)){ // find matching collection names
-						if (cName != props.applicationCollectionName && cName != props.applicationOutputsCollectionName){
-							impactResolver.addCollection(cName) // add collection of foreign application
-						}
-					}
-					else{
-						//if (props.verbose) println("$cName does not match pattern: $collectionMatcherPatterns")
-					}
-				}
-				// resolve external impacted files
-				def externalImpactedFiles = impactResolver.resolve()
-
-				// report scanning results
-				if (externalImpactedFiles.size()!=0) if (props.verbose) println("*** Identified external impacted files for changed file $changedFile")
-				externalImpactedFiles.each{ externalImpact ->
-					def Set<String> externalImpactList = collectionImpactsSetMap.get(externalImpact.getCollection()) ?: new HashSet<String>()
-					def impactRecord = "${externalImpact.getLname()} \t ${externalImpact.getFile()} \t ${externalImpact.getCollection()}"
-					// if (props.verbose) println("*** $impactRecord")
-					externalImpactList.add(impactRecord)
-					collectionImpactsSetMap.put(externalImpact.getCollection(), externalImpactList) // <collection,list of impacted files>
-				}
-			}
-			else {
-				println("*! build property reportExternalImpactsAnalysisDepths has in invalid value : ${props.reportExternalImpactsAnaylsisDepths} , valid: simple | deep")
-			}
-		}
-		else {
-			if (props.verbose) println("*** Analysis and reporting has been skipped for changed file $changedFile due to build framework configuration (see configuration of build property reportExternalImpactsAnalysisFileFilter)")
-		}
-	}
-
-	// generate reports by collection / application
-	collectionImpactsSetMap.each{ entry ->
-		externalImpactList = entry.value
-		if (externalImpactList.size()!=0){
-			// write impactedFiles per application to build workspace
-			String impactListFileLoc = "${props.buildOutDir}/externalImpacts_${entry.key}.${props.buildListFileExt}"
-			if (props.verbose) println("*** Writing report of external impacts to file $impactListFileLoc")
-			File impactListFile = new File(impactListFileLoc)
-			String enc = props.logEncoding ?: 'IBM-1047'
-			impactListFile.withWriter(enc) { writer ->
-				externalImpactList.each { file ->
-					// if (props.verbose) println file
-					writer.write("$file\n")
-				}
-			}
-		}
-	}
-
-}
-
-/*
- * Method to generate the Concurrent Changes reports and validate if the current build list intersects with concurrent changes
- */
-
-def generateConcurrentChangesReports(Set<String> buildList, Set<String> concurrentChangedFiles, Set<String> concurrentRenamedFiles, Set<String> concurrentDeletedFiles, String gitReference, RepositoryClient repositoryClient){
-	String concurrentChangesReportLoc = "${props.buildOutDir}/report_concurrentChanges.txt"
-
-	File concurrentChangesReportFile = new File(concurrentChangesReportLoc)
-	String enc = props.logEncoding ?: 'IBM-1047'
-	concurrentChangesReportFile.withWriterAppend(enc) { writer ->
-
-		if (!(concurrentChangedFiles.size() == 0 &&  concurrentRenamedFiles.size() == 0 && concurrentDeletedFiles.size() == 0)) {
-
-			if (props.verbose) println("** Writing report of concurrent changes to $concurrentChangesReportLoc for configuration $gitReference")
-
-			writer.write("\n=============================================== \n")
-			writer.write("** Report for configuration: $gitReference \n")
-			writer.write("========\n")
-
-			if (concurrentChangedFiles.size() != 0) {
-				writer.write("** Changed Files \n")
-				concurrentChangedFiles.each { file ->
-					if (props.verbose) println " Changed: ${file}"
-					if (buildList.contains(file)) {
-						writer.write("* $file is changed and intersects with the current build list.\n")
-						String msg = "*!! $file is changed on branch $gitReference and intersects with the current build list."
-						println msg
-						
-						// update build result
-						if (props.reportConcurrentChangesIntersectionFailsBuild && props.reportConcurrentChangesIntersectionFailsBuild.toBoolean()) {
-							props.error = "true"
-							buildUtils.updateBuildResult(errorMsg:msg,client:repositoryClient)
-						} else {
-							buildUtils.updateBuildResult(warningMsg:msg,client:repositoryClient)
-						}
-					}
-					else
-						writer.write("  $file\n")
-				}
-			}
-
-			if (concurrentRenamedFiles.size() != 0) {
-				writer.write("** Renamed Files \n")
-				concurrentRenamedFiles.each { file ->
-					if (props.verbose) println " Renamed: ${file}"
-					if (buildList.contains(file)) {
-						writer.write("* $file got renamed and intersects with the current build list.\n")
-						String msg = "*!! $file is renamed on branch $gitReference and intersects with the current build list."
-						println msg
-						
-						// update build result
-						if (props.reportConcurrentChangesIntersectionFailsBuild && props.reportConcurrentChangesIntersectionFailsBuild.toBoolean()) {
-							props.error = "true"
-							buildUtils.updateBuildResult(errorMsg:msg,client:repositoryClient)
-						} else {
-							buildUtils.updateBuildResult(warningMsg:msg,client:repositoryClient)
-						}
-					}
-					else
-						writer.write("  $file\n")
-				}
-			}
-
-			if (concurrentDeletedFiles.size() != 0) {
-				writer.write("** Deleted Files \n")
-				concurrentDeletedFiles.each { file ->
-					if (props.verbose) println " Deleted: ${file}"
-					if (buildList.contains(file)) {
-						writer.write("* $file is deleted and intersects with the current build list.\n")
-						String msg = "*!! $file is deleted on branch $gitReference and intersects with the current build list."
-						println msg
-						
-						// update build result
-						if (props.reportConcurrentChangesIntersectionFailsBuild && props.reportConcurrentChangesIntersectionFailsBuild.toBoolean()) {
-							props.error = "true"
-							buildUtils.updateBuildResult(errorMsg:msg,client:repositoryClient)
-						} else {
-							buildUtils.updateBuildResult(warningMsg:msg,client:repositoryClient)
-						}
-					}
-					else
-						writer.write("  $file\n")
-				}
 			}
 		}
 	}
